third_party/blink/web_tests/external/wpt/editing/other/typing-around-link-element-after-joining-paragraphs_action=Backspace-expected.txt

This is a testharness.js-based test.
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>" but got "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>" but got "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
  assert_equals: expected "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>" but got "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
  assert_equals: expected "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>" but got "<p><a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span></p>"
[FAIL] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span>" but got "<a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span>"
[FAIL] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a><b>XY</b><span style=\\"color:rgb(255, 0, 0)\\">bar</span>" but got "<a href=\\"about:blank\\" style=\\"color:rgb(0, 0, 255)\\"><b>foo</b></a>XY<span style=\\"color:rgb(255, 0, 0)\\">bar</span>"
Harness: the test ran to completion.
